shopsense-backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
google-generativeai>=0.5.0
pydantic>=2.7.0
python-dotenv>=1.0.1
duckduckgo-search>=5.0.0
